ui: updated package json versions of platform & ui for react & mantine
