+++ platform/firecamp-platform/package.json
@@ -46,7 +46,7 @@
   },
   "devDependencies": {
     "@types/chrome": "^0.0.204",
-    "@types/react-dom": "^17.0.13"
+    "@types/react-dom": "18.2.7"
   },
   "dependencies": {
     "@firecamp/agent-manager": "workspace:*",
@@ -62,7 +62,7 @@
     "@firecamp/url": "0.0.20",
     "@firecamp/utils": "^0.0.24",
     "@firecamp/websocket": "workspace:*",
-    "@mantine/form": "^6.0.17",
+    "@mantine/form": "^7.1.0",
     "@sentry/browser": "7.28.1",
     "@sentry/tracing": "7.28.1",
     "awesome-notifications": "3.1.3",
@@ -78,8 +78,8 @@
     "nanoid": "3.3.4",
     "prop-types": "^15.6.2",
     "re-resizable": "^4.7.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
     "react-error-boundary": "^3.1.4",
     "react-ga": "^3.1.2",
