+++ platform/firecamp-ui/package.json
@@ -59,8 +59,8 @@
     "@testing-library/react": "12.1.2",
     "@types/codemirror": "^5.60.5",
     "@types/markdown-it": "^12.2.3",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
     "@types/react-table": "^7.7.10",
     "autoprefixer": "^10.4.14",
     "babel-jest": "^29.1.2",
@@ -76,7 +76,8 @@
     "postcss": "^8.4.21",
     "postcss-cli": "^8.3.1",
     "raw-loader": "^4.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "sass": "^1.54.4",
     "sass-loader": "^13.0.2",
     "style-loader": "^3.3.1",
@@ -88,10 +89,10 @@
     "@firecamp/types": "^0.2.2",
     "@firecamp/ui": "workspace:*",
     "@firecamp/utils": "^0.0.24",
-    "@mantine/core": "^6.0.14",
-    "@mantine/form": "^6.0.17",
-    "@mantine/hooks": "^6.0.14",
-    "@mantine/modals": "^6.0.17",
+    "@mantine/core": "^7.1.0",
+    "@mantine/form": "^7.1.0",
+    "@mantine/hooks": "^7.1.0",
+    "@mantine/modals": "^7.1.0",
     "@monaco-editor/react": "4.4.5",
     "@radix-ui/react-dropdown-menu": "^2.0.1",
     "@radix-ui/react-scroll-area": "^1.0.2",
@@ -116,10 +117,6 @@
     "react-tiny-popover": "^7.0.1",
     "react-window": "^1.8.5",
     "vscode": "^1.1.37"
-  },
-  "peerDependencies": {
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
   },
   "babel": {
     "presets": [
