+++ platform/firecamp-ui/src/components/grid/RootContainer.tsx
@@ -4,7 +4,7 @@ import cx from 'classnames';
 
 // make sure to import css scripts in order 1. preflight, 2. mantine Core, 3. tailwind styles
 import "../../scss/preflight.css";
-// import "@mantine/core/styles.css";
+import "@mantine/core/styles.css";
 import "../../scss/tailwind.scss";
 
 import { IRootContainer } from "./interfaces/RootContainer.interfaces"
