+++ platform/firecamp-ui/src/components/theme/FirecampThemeProvider.interfaces.ts
@@ -1,14 +1,14 @@
 import {
   DefaultMantineColor,
   MantineProviderProps,
-  Tuple,
+  MantineColorsTuple,
 } from '@mantine/core';
 
 type ExtendedCustomColors = 'primary-color' | DefaultMantineColor;
 
 declare module '@mantine/core' {
   export interface MantineThemeColorsOverride {
-    colors: Record<ExtendedCustomColors, Tuple<string, 10>>;
+    colors: Record<ExtendedCustomColors, MantineColorsTuple>;
   }
 }
 export enum EFirecampThemeVariant {
@@ -17,5 +17,5 @@ export enum EFirecampThemeVariant {
   DarkPrimary = 'dark-primary',
   DarkSecondary = 'dark-secondary',
 }
-export type ColorType = Tuple<string, 10>;
+export type ColorType = MantineColorsTuple;
 export interface IFirecampThemeProvider extends MantineProviderProps {}
